chore: move utils functions to separate module
